Refactor timer service and I apply singleton
