+++ src/app/timer/timer.component.ts
@@ -1,71 +1,26 @@
 import { Component, OnInit, OnDestroy, Input, Output, EventEmitter } from '@angular/core';
+import { TimerService } from './timer.service';
 
 @Component({
   selector: 'app-timer',
   templateUrl: './timer.component.html',
-  styleUrls: ['./timer.component.scss']
+  styleUrls: ['./timer.component.scss'],
+  providers: [TimerService]
 })
 export class TimerComponent implements OnInit, OnDestroy {
-
-  private countdownTimerRef: any = null;
-  public countdown: number = 0;
-  public paused: boolean = true;
 
   @Input() init: number = 20;
   @Output() onComplete = new EventEmitter<void>();
 
-  constructor() { }
+  constructor(
+    public timer: TimerService
+  ) { }
 
   ngOnInit(): void {
-    if(this.init && this.init > 0) {
-      this.countdown = this.init;
-    }
+    this.timer.restartCountdown(this.init);
   }
 
-  ngOnDestroy(): void {
-    this.clearTimeout();
+  ngOnDestroy() {
+    this.timer.destroy();
   }
-
-  restartCountdown() {
-    if (this.init && this.init > 0) {
-      this.paused = true;
-      this.clearTimeout();
-      this.countdown = this.init;
-    }
-  }
-
-  toogleCountdown() {
-    this.paused = !this.paused;
-
-    if(this.paused == false) {
-      this.doCountdown();
-    }else {
-      this.clearTimeout();
-    }
-  }
-
-  private doCountdown() {
-    this.countdownTimerRef = setTimeout(() => {
-      this.countdown = this.countdown - 1;
-      this.processCountdown();
-    }, 1000);
-  }
-
-  private processCountdown() {
-    if (this.countdown == 0) {
-      this.onComplete.emit();
-      console.log("--countdown end--");
-    }
-    else {
-      this.doCountdown();
-    }
-  }
-
-  private clearTimeout() {
-    if (this.countdownTimerRef) {
-      clearTimeout(this.countdownTimerRef);
-      this.countdownTimerRef = null;
-    }
-  }
-
 }
